Key changes:
Removed conditional compilation of the different indices. Instead, now a parameter is used to identify which index should be used for current run.

Dllmain.cpp: A layer is added for creating DLL and entry point functions. This is used for calling from managed code. (It has also been used internally)